move nlx form to admin
