+++ src/openzaak/config/templates/admin/config_nlx.html
@@ -1,19 +1,28 @@
-{% extends "master.html" %}
-{% load i18n sniplates %}
+{% extends "admin/base_site.html" %}
+{% load i18n static sniplates %}
+
+{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static "admin/css/forms.css" %}">{% endblock %}
+
+{% block title %} {% trans "Service configuratie" %} {{ block.super }} {% endblock %}
+
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+<a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
+&rsaquo; <a href="{% url 'config:config-detail' %}">{% trans 'Service configuratie' %}</a>
+&rsaquo; {% trans 'NLX' %}
+</div>
+{% endblock %}
 
 {% block content %}
+<h1>{% trans 'Service configuratie: NLX' %}</h1>
+<h3>Step 1 of 2</h3>
 
-<div class="px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
-    <h1 class="display-4">Service configuratie: NLX</h1>
-    <p class="lead">Step 1 of 2</p>
-</div>
-
-<div class="container">
+<div id="content-main">
     <form method="post" enctype="multipart/form-data">
         {% csrf_token %}
         {% load_widgets form='snippets/forms.html' %}
 
-        <fieldset>
+        <fieldset class="module aligned">
             {{ form.non_field_errors }}
 
             {% form_field form.directory %}
@@ -21,15 +30,15 @@
 
         </fieldset>
 
-        <div class="btn-toolbar justify-content-between py-sm-3">
-            <a class="btn btn-danger" href="{% url 'config-detail' %}">{% trans "cancel" %}</a>
-            <div>
-                <a class="btn btn-outline-dark" href="{% url 'config-internal' %}">{% trans "skip" %}</a>
-                <input class="btn btn-success" type="submit" name="submit" value="{% trans "next" %}" />
-            </div>
+        <div class="submit-row">
+            <input class="default" type="submit" name="submit" value="{% trans "next" %}" />
+            <p class="deletelink-box">
+                <a class="deletelink" href="{% url 'config:config-detail' %}">{% trans "cancel" %}</a>
+            </p>
+            <input type="button" value="{% trans 'skip' %}" onclick="location.href='{% url 'config-internal' %}';">
         </div>
-    </form>
 
+    </form>
 </div>
 
 {% endblock %}
