+++ src/openzaak/config/templates/admin/nlx_inway.html
@@ -6,7 +6,6 @@
 {% block breadcrumbs %}
 <div class="breadcrumbs">
 <a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
-&rsaquo; <a href="#">{% trans 'Configuratie' %}</a>
 &rsaquo; {% trans 'NLX inway' %}
 </div>
 {% endblock %}
