+++ src/openzaak/config/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 
-from .views import InternalConfigView, NLXConfigView
+from .views import InternalConfigView
 
 urlpatterns = [
-    path("nlx/", NLXConfigView.as_view(), name="config-nlx"),
     path("internal/", InternalConfigView.as_view(), name="config-internal"),
 ]
