+++ src/openzaak/templates/snippets/forms.html
@@ -3,7 +3,7 @@
 
 How to render labels
 {% block _label %}
-{% if label %}<label for="{{ id_for_label }}" class="col-sm-2 col-form-label">{{ label }}</label>{% endif %}
+{% if label %}<label for="{{ id_for_label }}">{{ label }}</label>{% endif %}
 {% endblock %}
 
 How to render errors
@@ -18,39 +18,34 @@ How to render errors
 {% endblock %}
 
 
-
 {% block URLInput %}
-<div class="form-group row">
+<div class="form-row">
     {% reuse "_label" %}
-    <div class="col-sm-10">
-        {% with errors|yesno:'is-invalid,' as validation_class %}
-            {% reuse "input" type="url" css_classes='form-control '|add:validation_class %}
-        {% endwith %}
-        {% reuse "_errors" %}
-    </div>
+    {% with errors|yesno:'is-invalid,' as validation_class %}
+        {% reuse "input" type="url" css_classes='vURLField '|add:validation_class %}
+    {% endwith %}
+    {% reuse "_errors" %}
 </div>
 {% endblock %}
 
 
 {% block Select %}
-<div class="form-group row">
+<div class="form-row">
     {% reuse "_label" %}
-    <div class="col-sm-10">
-        <select name="{{ html_name }}" id="{{ id }}" {{ widget.attrs|flatattrs }} class="form-control {{ errors|yesno:'is-invalid,' }}">
-        {% for choice in choices %}
-            {% if choice.is_group %}
-            <optgroup label="{{ choice.value }}">
-                {% for val, display in choice.display %}
-                    {% reuse '_Select_Option' val=val value=value display=display %}
-                {% endfor %}
-            </optgroup>
-            {% else %}
-                {% reuse '_Select_Option' val=choice.value value=value display=choice.display %}
-            {% endif %}
-        {% endfor %}
-        </select>
-        {% reuse "_errors" %}
-    </div>
+    <select name="{{ html_name }}" id="{{ id }}" {{ widget.attrs|flatattrs }} class="form-control {{ errors|yesno:'is-invalid,' }}">
+    {% for choice in choices %}
+        {% if choice.is_group %}
+        <optgroup label="{{ choice.value }}">
+            {% for val, display in choice.display %}
+                {% reuse '_Select_Option' val=val value=value display=display %}
+            {% endfor %}
+        </optgroup>
+        {% else %}
+            {% reuse '_Select_Option' val=choice.value value=value display=choice.display %}
+        {% endif %}
+    {% endfor %}
+    </select>
+    {% reuse "_errors" %}
 </div>
 {% endblock %}
